usb: select usb-service rc files based on device type
i2c paths for 5.15 and 6.1 kernels are statically assigned,
so delete variations that do not exist from init rc file.
Depending on the device either i2c bus 6 or 11 are used, so
create permissions for both in separate init rc files.

Put the correct rc file to the vendor partition based on
target product.

Test: Manual test on device. Confirm that file reads related to
      file path error do not trigger error.
Bug: 318820040
Change-Id: I4cb67bcbba99ef32a5ddfb2487b92ee196563b90
Signed-off-by: RD Babiera <rdbabiera@google.com>