[v2] Disable MTE on certain system processes that have it enabled by default
A bunch of stuff in build/make/target/product/memtag-common.mk is specified has having MTE by default, in order to ensure that OEMs that have MTE-capable devices get a default level of MTE on security sensitive processes.

For now, on these devices, ensure that devices that enable MTE through the developer toggle (or the bootloader cmdline) only have the zygote with MTE.

Last attempt was accidentally only applied to dev boards (I thought the device product inherited), but now set it as the base of zumapro.

Bug: 327060622
Change-Id: I57b08b6edcf3ebdcbc790e29b89e70bc0edbc5fd
Test: Boot device, `fastboot oem mte enable`, follow instructions at b/b/327060622