shamp: Replace modem_svc with shamp
Bug: 328704096

Newer devices will depend upon the new soong build target for modem_svc
(now named shared_modem_platform). Since the target is included in
gs-common, and is thus on all devices, the only thing that needs to be
done it this build file is to specify that we need the lassen instance.

Test: build, flash and check for lassen logs
Doc: 328704096
Change-Id: I48af45ee76b3ec55422be81462212986b3d97322