Disable dumpstate strict-run flag in zumapro
The change disables the dumpstate strict-run flag in zumapro devices, which is enabled by default. Disabling this flag increases the timeout for capturing incident report to 120 seconds, from the default 20 seconds. This is done because 20s isn't enough to capture incident report proto in some cases, especially when device is throttling.

Bug: 329683355

Test: Booted a zumapro device with the change. Then manually thermal throttled the device and captured bugreports. In some tests Incident report took more than 20s, up to 55s, and in all these cases it was succesfully captured.

Change-Id: I0b12c041c8ac864897db53f44ce51292e1ca36bb